requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
python-decouple
django-cors-headers
psycopg2-binary
